requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiohttp_socks==0.10.1
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
attrs==25.3.0
av==14.4.0
azure-ai-documentintelligence==1.0.2
azure-core==1.34.0
azure-identity==1.23.0
beautifulsoup4==4.13.4
blinker==1.9.0
bottle==0.13.3
Brotli==1.1.0
browser-cookie3==0.20.1
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cobble==0.1.4
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.3
cssselect2==0.8.0
ctranslate2==4.4.0
curl_cffi==0.11.1
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.2.18
docopt==0.6.2
duckduckgo_search==8.0.2
EbookLib==0.19
einops==0.8.1
et_xmlfile==2.0.0
fastapi==0.115.12
faster-whisper==1.1.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.1
future==1.0.0
g4f==0.5.3.1
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.2
huggingface-hub==0.32.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
jeepney==0.9.0
Jinja2==3.1.6
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.8
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
lxml==5.4.0
lz4==4.4.4
magika==0.6.2
Mako==1.3.10
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.1.1
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
mss==10.0.0
multidict==6.4.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodriver==0.46.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
odfpy==1.4.1
olefile==0.47
omegaconf==2.3.0
onnxruntime==1.22.0
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pdfminer.six==20250506
pillow==11.2.1
platformdirs==4.3.8
plyer==2.1.0
primePy==1.3
primp==0.15.0
propcache==0.3.1
protobuf==4.25.7
proxy_tools==0.1.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
python-socks==2.7.1
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
pywebview==5.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.11
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.0
setuptools==80.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
speechbrain==1.0.3
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.7.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.0
torchmetrics==1.7.1
tqdm==4.67.1
transformers==4.52.3
triton==3.3.0
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
whisperx @ git+https://github.com/matheusbach/whisperx.git@dcf896a1a0bd54b0a55c8090bbd405a4d8cf7827
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.3
yarl==1.20.0
youtube-transcript-api==1.0.3
yt-dlp==2025.5.22
